More repo cleanup
